update order of the pages
also use the latest version of minima theme, which was released in 2019.

the master has a 3.0 going on but it is not released. it has a few
things i wanted, e.g remove the "subscribe from rss" in the home page.
but will wait and see to avoid too much customization.